Use the column ucod to count deaths
Using the column or the entire data row should not
make a difference. I made this change while trying to
solve the data issue that turned out to be related to
state code differences between years. (See a few commits
below.) I kept this change in, because that was the
final state of the processed data. (v.3.3)